requirements.txt
fastapi
uvicorn
gunicorn
uvloop
httptools
databases
asyncpg
python-dotenv
geopandas
psycopg2-binary